fix: refactor exceptions
