+++ src/main/java/io/supertokens/pluginInterface/webauthn/exceptions/UserIdNotFoundException.java
@@ -14,7 +14,8 @@
  *    under the License.
  */
 
-package io.supertokens.pluginInterface.webauthn;
+package io.supertokens.pluginInterface.webauthn.exceptions;
 
-public class DuplicateUserIdException extends Exception {
+public class UserIdNotFoundException extends Exception {
+
 }
